qcacmn: Change data type of modified_in_curr_session
Repetitive channel capture configurations usually support
16 groups to be confugured. To check if a particular group
is enabled by user, qdf_test_bit is being used here.

qdf_test_bit usually operates with uint32_t and it takes
the argument as unsigned long type. Since,
modified_in_curr_session is declared as uint16_t, and then
typecasted into unsigned long.

Change the data type of modified_in_curr_session
from uint16_t to 'unsigned long' to avoid compiler
issues.

CRs-Fixed: 2753267

+++ umac/cfr/dispatcher/inc/wlan_cfr_utils_api.h
@@ -444,7 +444,7 @@ struct cfr_rcc_param {
 	uint8_t num_grp_tlvs;
 
 	struct ta_ra_cfr_cfg curr[MAX_TA_RA_ENTRIES];
-	uint16_t modified_in_curr_session;
+	unsigned long modified_in_curr_session;
 	uint32_t capture_count            :16,
 		 capture_intval_mode_sel  :1,
 		 rsvd2                    :15;
@@ -626,7 +626,7 @@ QDF_STATUS wlan_cfr_pdev_close(struct wlan_objmgr_pdev *pdev);
  *
  * Return: No. of set bits
  */
-uint8_t count_set_bits(uint32_t value);
+uint8_t count_set_bits(unsigned long value);
 
 /**
  * wlan_cfr_is_feature_disabled() - Check if cfr feature is disabled
